added validation for required params
