+++ src/appmixer/googleSlides/core/DeleteSlide/DeleteSlide.js
@@ -4,6 +4,14 @@ module.exports = {
     async receive(context) {
         const { presentationId, slideId } = context.messages.in.content;
 
+        if (!presentationId) {
+            throw new context.CancelError('Presentation ID is required!');
+        }
+
+        if (!slideId) {
+            throw new context.CancelError('Slide ID is required!');
+        }
+
         // https://developers.google.com/slides/api/reference/rest/v1/presentations/batchUpdate
         const { data } = await context.httpRequest({
             method: 'POST',
