+++ src/appmixer/googleSlides/core/FindPresentations/FindPresentations.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const lib = require('../../lib.generated');
+const lib = require('../../lib');
 
 module.exports = {
     async receive(context) {
