+++ src/appmixer/googleSlides/lib.js
@@ -18,5 +18,110 @@ module.exports = {
         } else {
             throw new context.CancelError(`${fieldName} must be a string or an array`);
         }
+    },
+
+    async sendArrayOutput({ context, outputPortName = 'out', outputType = 'array', records = [] }) {
+        if (outputType === 'first') {
+            if (records.length === 0) {
+                throw new context.CancelError('No records available for first output type');
+            }
+            // One by one.
+            await context.sendJson(
+                { ...records[0], index: 0, count: records.length },
+                outputPortName
+            );
+        } else if (outputType === 'object') {
+            // One by one.
+            for (let index = 0; index < records.length; index++) {
+                await context.sendJson(
+                    { ...records[index], index, count: records.length },
+                    outputPortName
+                );
+            }
+        } else if (outputType === 'array') {
+            // All at once.
+            await context.sendJson({ result: records, count: records.length }, outputPortName);
+        } else if (outputType === 'file') {
+
+            // Into CSV file.
+            const csvString = toCsv(records);
+
+            let buffer = Buffer.from(csvString, 'utf8');
+            const componentName = context.flowDescriptor[context.componentId].label || context.componentId;
+            const fileName = `${context.config.outputFilePrefix || DEFAULT_PREFIX}-${componentName}.csv`;
+            const savedFile = await context.saveFileStream(pathModule.normalize(fileName), buffer);
+
+            await context.log({ step: 'File was saved', fileName, fileId: savedFile.fileId });
+            await context.sendJson({ fileId: savedFile.fileId }, outputPortName);
+        } else {
+            throw new context.CancelError('Unsupported outputType ' + outputType);
+        }
+    },
+
+    getProperty(obj, path) {
+        return path.split('.').reduce((acc, part) => acc?.[part], obj);
+    },
+
+    getOutputPortOptions(context, outputType, itemSchema, { label, value }) {
+
+        if (outputType === 'object' || outputType === 'first') {
+            const options = Object.keys(itemSchema)
+                .reduce((res, field) => {
+                    const schema = itemSchema[field];
+                    const { title: label, ...schemaWithoutTitle } = schema;
+
+                    res.push({
+                        label, value: field, schema: schemaWithoutTitle
+                    });
+                    return res;
+                }, [{
+                    label: 'Current Item Index',
+                    value: 'index',
+                    schema: { type: 'integer' }
+                }, {
+                    label: 'Items Count',
+                    value: 'count',
+                    schema: { type: 'integer' }
+                }]);
+
+            return context.sendJson(options, 'out');
+        }
+
+        if (outputType === 'array') {
+            return context.sendJson([{
+                label,
+                value,
+                schema: {
+                    type: 'array',
+                    items: { type: 'object', properties: itemSchema }
+                }
+            }], 'out');
+        }
+
+        if (outputType === 'file') {
+            return context.sendJson([{ label: 'File ID', value: 'fileId' }], 'out');
+        }
     }
+};
+
+/**
+ * @param {array} array
+ * @returns {string}
+ */
+const toCsv = (array) => {
+    const headers = Object.keys(array[0]);
+
+    return [
+        headers.join(','),
+
+        ...array.map(items => {
+            return Object.values(items).map(property => {
+                if (typeof property === 'object') {
+                    return JSON.stringify(property);
+                }
+                return property;
+            }).join(',');
+        })
+
+    ].join('\n');
 };
